+++ litex/src/define_algorithm_stmt.rs
@@ -8,18 +8,18 @@ pub struct DefineAlgorithmStmt {
     pub name: String,
     pub params: Vec<String>,
     pub return_or_algo_if: Vec<AlgoReturnOrAlgoIf>,
-    pub line_file_index: Option<(u16, usize)>,
+    pub line_file_index: Option<(usize, usize)>,
 }
 
 pub struct AlgoReturn {
     pub value: Obj,
-    pub line_file_index: Option<(u16, usize)>,
+    pub line_file_index: Option<(usize, usize)>,
 }
 
 pub struct AlgoIf {
     pub condition: AndFactOrSpecFact,
     pub return_stmt: AlgoReturn,
-    pub line_file_index: Option<(u16, usize)>,
+    pub line_file_index: Option<(usize, usize)>,
 }
 
 pub enum AlgoReturnOrAlgoIf {
@@ -28,7 +28,7 @@ pub enum AlgoReturnOrAlgoIf {
 }
 
 impl DefineAlgorithmStmt {
-    pub fn new(name: String, params: Vec<String>, return_or_algo_if: Vec<AlgoReturnOrAlgoIf>, line_file_index: Option<(u16, usize)>) -> Self {
+    pub fn new(name: String, params: Vec<String>, return_or_algo_if: Vec<AlgoReturnOrAlgoIf>, line_file_index: Option<(usize, usize)>) -> Self {
         DefineAlgorithmStmt { name, params, return_or_algo_if, line_file_index }
     }
 }
@@ -69,19 +69,19 @@ impl fmt::Display for DefineAlgorithmStmt {
 }
 
 impl AlgoReturn {
-    pub fn new(value: Obj, line_file_index: Option<(u16, usize)>) -> Self {
+    pub fn new(value: Obj, line_file_index: Option<(usize, usize)>) -> Self {
         AlgoReturn { value, line_file_index }
     }
 }
 
 impl AlgoIf {
-    pub fn new(condition: AndFactOrSpecFact, return_stmt: AlgoReturn, line_file_index: Option<(u16, usize)>) -> Self {
+    pub fn new(condition: AndFactOrSpecFact, return_stmt: AlgoReturn, line_file_index: Option<(usize, usize)>) -> Self {
         AlgoIf { condition, return_stmt, line_file_index }
     }
 }
 
 impl AlgoReturnOrAlgoIf {
-    pub fn line_file(&self) -> Option<(u16, usize)> {
+    pub fn line_file(&self) -> Option<(usize, usize)> {
         match self {
             AlgoReturnOrAlgoIf::AlgoReturn(algo_return) => algo_return.line_file_index,
             AlgoReturnOrAlgoIf::AlgoIf(algo_if) => algo_if.line_file_index,
