+++ litex/src/know_stmt.rs
@@ -3,15 +3,15 @@ use crate::fact::Fact;
 
 pub struct KnowStmt {
     pub facts: Vec<Fact>,
-    pub line_file_index: Option<(u16, usize)>,
+    pub line_file_index: Option<(usize, usize)>,
 }
 
 impl KnowStmt {
-    pub fn new(facts: Vec<Fact>, line_file_index: Option<(u16, usize)>) -> Self {
+    pub fn new(facts: Vec<Fact>, line_file_index: Option<(usize, usize)>) -> Self {
         KnowStmt { facts, line_file_index }
     }
 
-    pub fn line_file(&self) -> Option<(u16, usize)> {
+    pub fn line_file(&self) -> Option<(usize, usize)> {
         self.line_file_index
     }
 }
