+++ litex/src/prove_by_builtin_techniques_stmt.rs
@@ -21,7 +21,7 @@ pub enum ProveByBuiltinTechniqueStmt {
 // f $in fn(a A, b B: $p(a, b)) C {$q(a, b, f)} => forall a A, b B: $p(a, b) => ((a, b), f(a, b)) $in f; forall x f: exist a A, b B st {$p(a, b), ((a, b), f(a, b)) = x}
 pub struct ViewFnAsSetStmt {
     pub function: Obj,
-    pub line_file_index: Option<(u16, usize)>,
+    pub line_file_index: Option<(usize, usize)>,
 }
 
 
@@ -29,7 +29,7 @@ pub struct ProveEqualSetByDefStmt {
     pub left: Obj,
     pub right: Obj,
     pub proof: Vec<Stmt>,
-    pub line_file_index: Option<(u16, usize)>,
+    pub line_file_index: Option<(usize, usize)>,
 }
 
 
@@ -44,7 +44,7 @@ pub struct ProveForStmt {
     pub dom_facts: Vec<OrFactOrAndFactOrSpecFact>,
     pub then_facts: Vec<OrFactOrAndFactOrSpecFact>,
     pub proof: Vec<Stmt>,
-    pub line_file_index: Option<(u16, usize)>,
+    pub line_file_index: Option<(usize, usize)>,
 }
 
 pub struct ProveByInductionStmt {
@@ -52,7 +52,7 @@ pub struct ProveByInductionStmt {
     pub param: String,
     pub proof: Vec<Stmt>,
     pub induc_from: Obj,
-    pub line_file_index: Option<(u16, usize)>,
+    pub line_file_index: Option<(usize, usize)>,
 }
 
 pub struct ProveByEnumerationStmt {
@@ -60,7 +60,7 @@ pub struct ProveByEnumerationStmt {
     pub param_sets: Vec<Obj>,
     pub to_prove: Vec<Fact>,
     pub proof: Vec<Stmt>,
-    pub line_file_index: Option<(u16, usize)>,
+    pub line_file_index: Option<(usize, usize)>,
 }
 
 pub struct ProveCaseByCaseStmt {
@@ -68,25 +68,25 @@ pub struct ProveCaseByCaseStmt {
     pub then_facts: Vec<Fact>,
     pub proofs: Vec<Vec<Stmt>>,
     pub impossible_facts: Vec<Option<OrFactOrAndFactOrSpecFact>>,
-    pub line_file_index: Option<(u16, usize)>,
+    pub line_file_index: Option<(usize, usize)>,
 }
 
 pub struct ProveByContradictionStmt {
     pub to_prove: Fact,
     pub proof: Vec<Stmt>,
     pub impossible_fact: OrFactOrAndFactOrSpecFact,
-    pub line_file_index: Option<(u16, usize)>,
+    pub line_file_index: Option<(usize, usize)>,
 }
 
 
 impl ProveByEnumerationStmt {
-    pub fn new(params: Vec<String>, param_sets: Vec<Obj>, to_prove: Vec<Fact>, proof: Vec<Stmt>, line_file_index: Option<(u16, usize)>) -> Self {
+    pub fn new(params: Vec<String>, param_sets: Vec<Obj>, to_prove: Vec<Fact>, proof: Vec<Stmt>, line_file_index: Option<(usize, usize)>) -> Self {
         ProveByEnumerationStmt { params, param_sets, to_prove, proof, line_file_index }
     }
 }
 
 impl ProveCaseByCaseStmt {
-    pub fn new(cases: Vec<AndFactOrSpecFact>, then_facts: Vec<Fact>, proofs: Vec<Vec<Stmt>>, impossible_facts: Vec<Option<OrFactOrAndFactOrSpecFact>>, line_file_index: Option<(u16, usize)>) -> Self {
+    pub fn new(cases: Vec<AndFactOrSpecFact>, then_facts: Vec<Fact>, proofs: Vec<Vec<Stmt>>, impossible_facts: Vec<Option<OrFactOrAndFactOrSpecFact>>, line_file_index: Option<(usize, usize)>) -> Self {
         ProveCaseByCaseStmt { cases, then_facts, proofs, impossible_facts, line_file_index }
     }
 }
@@ -106,7 +106,7 @@ impl fmt::Display for ProveCaseByCaseStmt {
 }
 
 impl ProveByContradictionStmt {
-    pub fn new(to_prove: Fact, proof: Vec<Stmt>, impossible_fact: OrFactOrAndFactOrSpecFact, line_file_index: Option<(u16, usize)>) -> Self {
+    pub fn new(to_prove: Fact, proof: Vec<Stmt>, impossible_fact: OrFactOrAndFactOrSpecFact, line_file_index: Option<(usize, usize)>) -> Self {
         ProveByContradictionStmt { to_prove, proof, impossible_fact, line_file_index }
     }
 }
@@ -132,7 +132,7 @@ impl fmt::Display for ProveByBuiltinTechniqueStmt {
 }
 
 impl ProveByBuiltinTechniqueStmt {
-    pub fn line_file(&self) -> Option<(u16, usize)> {
+    pub fn line_file(&self) -> Option<(usize, usize)> {
         match self {
             ProveByBuiltinTechniqueStmt::ProveCaseByCase(prove_case_by_case) => prove_case_by_case.line_file_index,
             ProveByBuiltinTechniqueStmt::ProveByContradiction(claim_prove_by_contradiction_stmt) => claim_prove_by_contradiction_stmt.line_file_index,
@@ -152,7 +152,7 @@ impl fmt::Display for ProveByEnumerationStmt {
 }
 
 impl ProveByInductionStmt {
-    pub fn new(fact: Vec<OrFactOrAndFactOrSpecFact>, param: String, proof: Vec<Stmt>, induc_from: Obj, line_file_index: Option<(u16, usize)>) -> Self {
+    pub fn new(fact: Vec<OrFactOrAndFactOrSpecFact>, param: String, proof: Vec<Stmt>, induc_from: Obj, line_file_index: Option<(usize, usize)>) -> Self {
         ProveByInductionStmt { fact, param, proof, induc_from, line_file_index }
     }
 }
@@ -178,7 +178,7 @@ impl fmt::Display for ProveForStmt {
 }
 
 impl ProveForStmt {
-    pub fn new(params: Vec<String>, param_sets: Vec<ClosedRangeOrRange>, dom_facts: Vec<OrFactOrAndFactOrSpecFact>, then_facts: Vec<OrFactOrAndFactOrSpecFact>, proof: Vec<Stmt>, line_file_index: Option<(u16, usize)>) -> Self {
+    pub fn new(params: Vec<String>, param_sets: Vec<ClosedRangeOrRange>, dom_facts: Vec<OrFactOrAndFactOrSpecFact>, then_facts: Vec<OrFactOrAndFactOrSpecFact>, proof: Vec<Stmt>, line_file_index: Option<(usize, usize)>) -> Self {
         ProveForStmt { params, param_sets, dom_facts, then_facts, proof, line_file_index }
     }
 }
@@ -202,7 +202,7 @@ impl fmt::Display for ProveEqualSetByDefStmt {
 }
 
 impl ProveEqualSetByDefStmt {
-    pub fn new(left: Obj, right: Obj, proof: Vec<Stmt>, line_file_index: Option<(u16, usize)>) -> Self {
+    pub fn new(left: Obj, right: Obj, proof: Vec<Stmt>, line_file_index: Option<(usize, usize)>) -> Self {
         ProveEqualSetByDefStmt { left, right, proof, line_file_index }
     }
 }
@@ -214,7 +214,7 @@ impl fmt::Display for ViewFnAsSetStmt {
 }
 
 impl ViewFnAsSetStmt {
-    pub fn new(function: Obj, line_file_index: Option<(u16, usize)>) -> Self {
+    pub fn new(function: Obj, line_file_index: Option<(usize, usize)>) -> Self {
         ViewFnAsSetStmt { function, line_file_index }
     }
 }
