+++ litex/src/prove_stmt.rs
@@ -4,11 +4,11 @@ use crate::helper::vec_to_string_add_four_spaces_at_beginning_of_each_line;
 
 pub struct ProveStmt {
     pub proof: Vec<Stmt>,
-    pub line_file_index: Option<(u16, usize)>,
+    pub line_file_index: Option<(usize, usize)>,
 }
 
 impl ProveStmt {
-    pub fn new(proof: Vec<Stmt>, line_file_index: Option<(u16, usize)>) -> Self {
+    pub fn new(proof: Vec<Stmt>, line_file_index: Option<(usize, usize)>) -> Self {
         ProveStmt { proof, line_file_index }
     }
 }
